+++ codebase/apps/mdv_utils/src/MdvMultFilt/Process.cc
@@ -334,7 +334,7 @@ void Process::_initialize_out(DsMdvx &New, Params *P)
   // information, such as the setup parameters for a radar.
   // Here, we just pass them on from input to output.
   //
-  for (int ii = 0; ii < New.getNChunks(); ii++) {
+  for (size_t ii = 0; ii < New.getNChunks(); ii++) {
     MdvxChunk *chunk = new MdvxChunk(*New.getChunkByNum(ii));
     Out.addChunk(chunk);
   }
